Push check to function in utils
