+++ python/cusignal/test/test_peak_finding.py
@@ -15,23 +15,11 @@
 # import cusignal
 # import numpy as np
 # import pytest
-# import pytest_benchmark
 
-# from cusignal.test.utils import array_equal
+# from cusignal.test.utils import array_equal, _check_rapids_pytest_benchmark
 # from scipy import signal
 
-# try:
-#     from rapids_pytest_benchmark import setFixtureParamNames
-# except ImportError:
-#     print("\n\nWARNING: rapids_pytest_benchmark is not installed, "
-#           "falling back to pytest_benchmark fixtures.\n")
-
-#     # if rapids_pytest_benchmark is not available, just perfrom time-only
-#     # benchmarking and replace the util functions with nops
-#     gpubenchmark = pytest_benchmark.plugin.benchmark
-
-#     def setFixtureParamNames(*args, **kwargs):
-#         pass
+# gpubenchmark = _check_rapids_pytest_benchmark()
 
 # # Missing
 # # argrelmin
