+++ python/cusignal/test/utils.py
@@ -12,8 +12,28 @@
 # limitations under the License.
 
 import numpy as np
+import pytest_benchmark
 
 
 def array_equal(a, b, tol=1e-4, with_sign=True):
     res = np.max(np.abs(a - b)) < tol
     return res
+
+
+def _check_rapids_pytest_benchmark():
+    try:
+        from rapids_pytest_benchmark import setFixtureParamNames
+    except ImportError:
+        print(
+            "\n\nWARNING: rapids_pytest_benchmark is not installed, "
+            "falling back to pytest_benchmark fixtures.\n"
+        )
+
+        # if rapids_pytest_benchmark is not available, just perfrom time-only
+        # benchmarking and replace the util functions with nops
+        gpubenchmark = pytest_benchmark.plugin.benchmark
+
+        def setFixtureParamNames(*args, **kwargs):
+            pass
+
+        return gpubenchmark
